Refactor exporters to separate vocabulary exporters from base
