+++ ruby/lib/goi/exporter.rb
@@ -1,7 +1,3 @@
-require_relative 'exporter/config'
-require_relative 'exporter/base_exporter'
-require_relative 'exporter/io_exporter'
-require_relative 'exporter/google_sheet_exporter'
-require_relative 'exporter/anki_exporter'
-require_relative 'exporter/sql_file_exporter'
-require_relative 'exporter/sequel_exporter'
+require_relative 'exporter/core'
+require_relative 'exporter/vocabulary'
+require_relative 'exporter/grammar'
