+++ ruby/lib/goi/exporter/core.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require_relative 'core/config'
+require_relative 'core/base_exporter'
+
+require_relative 'core/anki_exporter'
