+++ ruby/lib/goi/exporter/core/config.rb
@@ -1,0 +1,16 @@
+module Goi
+  module Exporter
+    module Core
+      class Config
+
+        def initialize(file_pathname: nil, db_config: nil)
+          @file_pathname = file_pathname
+          @db_config = db_config
+        end
+
+        attr_reader :file_pathname, :db_config
+
+      end
+    end
+  end
+end
