+++ ruby/lib/goi/exporter/grammar.rb
@@ -1,0 +1,2 @@
+# frozen_string_literal: true
+
