+++ ruby/lib/goi/exporter/vocabulary.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require_relative 'vocabulary/anki_exporter'
+require_relative 'vocabulary/google_sheet_exporter'
+require_relative 'vocabulary/io_exporter'
+require_relative 'vocabulary/sequel_exporter'
+require_relative 'vocabulary/sql_file_exporter'
