--- ruby/sig/lib/goi/exporter/anki_exporter.rbs
@@ -1,78 +1,0 @@
-module Goi
-  module Exporter
-
-    class BaseAnkiExporter < BaseExporter
-
-      def initialize: (config: Config) -> void
-
-      attr_reader config: Config
-
-      def export: (linkages: [Model::Vocabulary::Linkages] linkages) -> void
-
-      private
-
-      def deck: () -> String?
-
-      def note_type: () -> String?
-
-      def header_row: () -> [String]?
-
-      def tags_column_index: () -> Integer?
-
-      def note_id: (linkage: Model::Vocabulary::Linkages) -> String?
-
-      def linkage_rows: (linkages: [Model::Vocabulary::Linkages] linkages) -> [[String?]]
-
-      def linkage_row: (linkage: Model::Vocabulary::Linkages linkage) -> [String?]?
-
-      def to_array_field: ([untyped] array) -> String?
-
-      def humanize_const: (String s) -> String
-
-      def tagize: (String s) -> String
-
-    end
-
-    class BaseAnkiVocabExporter < BaseAnkiExporter
-
-      NOTE_ID_HEADER: String
-
-      TAGS_HEADER: String
-
-      VOCABULARY_HEADERS: [String]
-
-      CONJUGATION_HEADERS: [String]
-
-      private
-
-      def note_id_fields: (Model::Vocabulary::Linkages linkage) -> [String]
-
-      def tags_fields: (Model::Vocabulary::Linkages linkage) -> [String]
-
-      def vocabulary_fields: (Model::Vocabulary::Linkages linkage) -> [String?]
-
-      def conjugation_fields: (Model::Vocabulary::ConjugationSet conjugation_set) -> [String?]
-
-      def phonetic_spelling_field: (Model::Vocabulary::Linkages linkage) -> String
-
-      def alt_spelling_field: (Model::Vocabulary::Linkages linkage) -> String?
-
-      def tags_field: (Model::Vocabulary::Vocabulary vocabulary) -> String?
-
-    end
-
-    class AnkiVocabExporter < BaseAnkiVocabExporter
-    end
-
-    class AnkiConjugationExporter < BaseAnkiVocabExporter
-
-      private
-
-      def vocabulary_row: (linkage: Model::Vocabulary::Linkages linkage) -> [String]
-
-      def conjugation_row: (conjugation_set: Model::Vocabulary::ConjugationSet conjugation_set) -> [String]
-
-    end
-
-  end
-end
