--- ruby/sig/lib/goi/exporter/base_exporter.rbs
@@ -1,15 +1,0 @@
-module Goi
-  module Exporter
-    class BaseExporter
-      def initialize: (config: Goi::Exporter::Config) -> void
-
-      attr_reader config: Config
-
-      def export: (linkages: [Model::Vocabulary::Linkages] linkages) -> void
-
-      private
-
-      def config_out_open: (){(IO) -> void} -> void
-    end
-  end
-end
