--- ruby/sig/lib/goi/exporter/config.rbs
@@ -1,15 +1,0 @@
-module Goi
-  module Exporter
-    class Config
-
-      type db_config = {database: String, user: String, password: String, host: String}
-
-      def initialize: (file_pathname: Pathname?, db_config: db_config?) -> void
-
-      attr_reader file_pathname: Pathname?
-
-      attr_reader db_config: db_config?
-
-    end
-  end
-end
