+++ ruby/sig/lib/goi/exporter/core/base_exporter.rbs
@@ -1,0 +1,17 @@
+module Goi
+  module Exporter
+    module Core
+      class BaseExporter
+        def initialize: (config: Goi::Exporter::Core::Config) -> void
+
+        attr_reader config: Config
+
+        def export: (linkages: [Model::Vocabulary::Linkages] linkages) -> void
+
+        private
+
+        def config_out_open: () { (IO) -> void } -> void
+      end
+    end
+  end
+end
