+++ ruby/sig/lib/goi/exporter/core/config.rbs
@@ -1,0 +1,17 @@
+module Goi
+  module Exporter
+    module Core
+      class Config
+
+        type db_config = { database: String, user: String, password: String, host: String }
+
+        def initialize: (file_pathname: Pathname?, db_config: db_config?) -> void
+
+        attr_reader file_pathname: Pathname?
+
+        attr_reader db_config: db_config?
+
+      end
+    end
+  end
+end
