--- ruby/sig/lib/goi/exporter/io_exporter.rbs
@@ -1,11 +1,0 @@
-module Goi
-  module Exporter
-    class IOExporter < BaseExporter
-      def initialize: (config: Config) -> void
-
-      attr_reader config: Config
-
-      def export: (linkages: [Model::Vocabulary::Linkages] linkages) -> void
-    end
-  end
-end
