--- ruby/sig/lib/goi/exporter/sql_file_exporter.rbs
@@ -1,19 +1,0 @@
-module Goi
-  module Exporter
-    class SqlFileExporter
-
-      def initialize: (config: Config) -> void
-
-      attr_reader config: Config
-
-      attr_reader db: untyped
-
-      private
-
-      attr_reader record_builder: Goi::SQL::VocabularyRecordBuilder
-
-      def export: (linkages: [Model::Vocabulary::Linkages] linkages) -> untyped
-
-    end
-  end
-end
