+++ ruby/sig/lib/goi/exporter/vocabulary/anki_exporter.rbs
@@ -1,0 +1,48 @@
+module Goi
+  module Exporter
+    module Vocabulary
+
+      class BaseAnkiVocabExporter < Goi::Exporter::Core::BaseAnkiExporter
+
+        NOTE_ID_HEADER: String
+
+        TAGS_HEADER: String
+
+        VOCABULARY_HEADERS: [String]
+
+        CONJUGATION_HEADERS: [String]
+
+        private
+
+        def note_id_fields: (Model::Vocabulary::Linkages linkage) -> [String]
+
+        def tags_fields: (Model::Vocabulary::Linkages linkage) -> [String]
+
+        def vocabulary_fields: (Model::Vocabulary::Linkages linkage) -> [String?]
+
+        def conjugation_fields: (Model::Vocabulary::ConjugationSet conjugation_set) -> [String?]
+
+        def phonetic_spelling_field: (Model::Vocabulary::Linkages linkage) -> String
+
+        def alt_spelling_field: (Model::Vocabulary::Linkages linkage) -> String?
+
+        def tags_field: (Model::Vocabulary::Vocabulary vocabulary) -> String?
+
+      end
+
+      class AnkiVocabExporter < BaseAnkiVocabExporter
+      end
+
+      class AnkiConjugationExporter < BaseAnkiVocabExporter
+
+        private
+
+        def vocabulary_row: (linkage: Model::Vocabulary::Linkages linkage) -> [String]
+
+        def conjugation_row: (conjugation_set: Model::Vocabulary::ConjugationSet conjugation_set) -> [String]
+
+      end
+
+    end
+  end
+end
