+++ ruby/sig/lib/goi/exporter/vocabulary/sequel_exporter.rbs
@@ -1,0 +1,25 @@
+module Goi
+  module Exporter
+    module Vocabulary
+
+      class SequelExporter < Goi::Exporter::Core::BaseExporter
+
+        private
+
+        attr_reader record_builder: Goi::SQL::VocabularyRecordBuilder
+
+        def db: () -> untyped
+
+        def naive_export: (linkages: [Model::Vocabulary::Linkages]) -> void
+
+        def delete_all: () -> void
+
+        def write_linkages: (linkages: [Model::Vocabulary::Linkages]) -> void
+
+        def write_record_group: (record_group: SQL::VocabularyRecordBuilder::record_group) -> void
+
+      end
+
+    end
+  end
+end
